refactor: move const values into shared utils file
