+++ src/dataExplorer/context/measurements.tsx
@@ -12,12 +12,6 @@ import {
 } from 'src/shared/constants/queryBuilder'
 
 // Contexts
-import {
-  IMPORT_REGEXP,
-  IMPORT_INFLUX_SCHEMA,
-  SAMPLE_DATA_SET,
-  FROM_BUCKET,
-} from 'src/dataExplorer/context/newDataExplorer'
 import {QueryContext} from 'src/shared/contexts/query'
 
 // Types
@@ -25,6 +19,12 @@ import {QueryScope, RemoteDataState} from 'src/types'
 
 // Utils
 import {isFlagEnabled} from 'src/shared/utils/featureFlag'
+import {
+  IMPORT_REGEXP,
+  IMPORT_INFLUX_SCHEMA,
+  SAMPLE_DATA_SET,
+  FROM_BUCKET,
+} from 'src/dataExplorer/shared/utils'
 
 interface MeasurementContextType {
   measurements: string[]
