+++ src/dataExplorer/context/newDataExplorer.tsx
@@ -8,14 +8,6 @@ import {TagsContext} from 'src/dataExplorer/context/tags'
 // Types
 import {Bucket} from 'src/types'
 
-export const IMPORT_REGEXP = 'import "regexp"\n'
-export const IMPORT_INFLUX_SCHEMA = 'import "influxdata/influxdb/schema"'
-export const SAMPLE_DATA_SET = (bucketID: string) =>
-  `import "influxdata/influxdb/sample"\nsample.data(set: "${bucketID}")`
-export const FROM_BUCKET = (bucketName: string) =>
-  `from(bucket: "${bucketName}")`
-
-export const LOCAL_LIMIT = 8
 interface NewDataExplorerContextType {
   // Schema
   selectedBucket: Bucket
